+++ SemestralnaPraca_AUS_02_with_GUI/App.h
@@ -22,7 +22,7 @@ public:
 
 	structures::ArrayList<Area*>* getAreasNazov(std::string nazov);
 	structures::ArrayList<Area*>* getAreasVolici(int pocetOd, int pocetDo, int kolo);
-	structures::ArrayList<Area*>* getAreasUcast(int ucastOd, int ucastDo, int kolo);
+	structures::ArrayList<Area*>* getAreasUcast(double ucastOd, double ucastDo, int kolo);
 
 	structures::SortedSequenceTable<string, Okres*>* getOkresy() { return this->okresySorted; }
 	structures::SortedSequenceTable<string, Obec*>* getObce() { return this->obceSorted; }
